+++ src/Scriban.Tests/TestParser.cs
@@ -28,6 +28,31 @@ namespace Scriban.Tests
         private const string BuiltinMarkdownDocFile = @"..\..\..\..\..\doc\builtins.md";
 
         [Test]
+        public void TestMemberDot()
+        {
+            var input = @"{{ a?.b.c }}";
+            var template = Template.Parse(input);
+            var result = template.Render();
+            Assert.AreEqual("", result);
+        }
+
+        [Test]
+        public void TestFailingError()
+        {
+            var input = @"{{ 
+  for $s in Foo
+      {{ if $s
+          false
+      else if $s
+          false
+      end
+  end 
+}}";
+            var template = Template.Parse(input);
+            Assert.True(template.HasErrors);
+        }
+
+        [Test]
         public void TestRoundtrip()
         {
             var text = "This is a text {{ code # With some comment }} and a text";
@@ -517,6 +542,52 @@ end
 
             Console.WriteLine(result);
         }
+
+        [Test]
+        public void TestIndent()
+        {
+            var input = @"{{ a_multi_line_value = ""test1\ntest2\ntest3\n"" ~}}
+   {{ a_multi_line_value }}Hello
+";
+            var template = Template.Parse(input);
+            var result = template.Render();
+            result = TextAssert.Normalize(result);
+
+            TextAssert.AreEqual(TextAssert.Normalize(@"   test1
+   test2
+   test3
+Hello
+"), result);
+        }
+
+        [Test]
+        public void TestIndentSkippedWithGreedyOnPreviousLine()
+        {
+            var input = @"{{ a_multi_line_value = ""test1\ntest2\ntest3\n"" -}}
+   {{ a_multi_line_value }}Hello
+";
+            var template = Template.Parse(input);
+            var result = template.Render();
+            result = TextAssert.Normalize(result);
+
+            TextAssert.AreEqual(TextAssert.Normalize(@"test1
+test2
+test3
+Hello
+"), result);
+        }
+
+        [Test]
+        public void TestIndent2()
+        {
+            var input = @"  {{data}}";
+            var template = Template.Parse(input);
+            var result = template.Render(new { data = "test\ntest2" });
+            result = TextAssert.Normalize(result);
+
+            TextAssert.AreEqual("  test\n  test2", result);
+        }
+
 
         [TestCaseSource("ListTestFiles", new object[] { "000-basic" })]
         public static void A000_basic(string inputName)