+++ src/Scriban.Tests/TestRuntime.cs
@@ -19,9 +19,106 @@ using Scriban.Syntax;
 
 namespace Scriban.Tests
 {
+    delegate string Args(object[] args);
+
     [TestFixture]
     public class TestRuntime
     {
+        [Test]
+        public void TestEnumerator()
+        {
+            var input = @"{{
+  queue.add 'a'
+  for x in queue.flush
+    x
+    if x == 'a'; queue.add 'b'; end
+  end
+}}";
+            var template = Template.Parse(input);
+
+            var test = template.Render(new { queue = new QueueBuiltin() });
+            Assert.AreEqual("ab", test);
+        }
+
+        class QueueBuiltin : ScriptObject
+        {
+            static Queue<string> queue = new();
+
+            public static void Add(string x) => queue.Enqueue(x);
+
+            public static IEnumerable<string> Flush()
+            {
+                while (queue.TryDequeue(out var x))
+                    yield return x;
+            }
+        }
+
+
+        [Test]
+        public void TestDateParse()
+        {
+            Template template = Template.Parse(
+                @"{{date.format='%FT%T.%N%Z'}}{{ date.parse '2018~06~17~13~59~+08:00' '%Y~%m~%d~%H~%M~%Z' }}");
+            var result = template.Render();
+            Console.WriteLine(result);
+        }
+
+
+        [Test]
+        public void TestLoop()
+        {
+
+            var template = Template.Parse(@"{{
+my_function(x) = x * i
+result = 0
+for i in [1,2,3,4]
+    result = result + (my_function 10)
+end
+}}Result: {{ result }}
+");
+            var result = template.Render();
+
+
+
+
+        }
+
+
+
+
+
+        [Test]
+        public void TestPars()
+        {
+            string Dump(params object[] args)
+            {
+                return "hello";
+            }
+
+            ScriptObject model = new ScriptObject();
+            ScriptObject debug = new ScriptObject();
+            Args dump = Dump;
+
+            debug.Import("dump", dump);
+            model["debug"] = debug;
+
+            var input = "{{debug.dump(10, \"hello\", [0, 1, 2])}}";
+            var template = Template.Parse(input);
+            var result = template.Render(model);
+
+            Assert.AreEqual("hello", result);
+        }
+
+        [Test]
+        public void TestUlong()
+        {
+            var input = @"{{if value > 0; 1; else; 2; end;}}";
+
+            var template = Template.Parse(input);
+            var result = template.Render(new { value = (ulong) 1} );
+            Assert.AreEqual("1", result);
+        }
+
         [Test]
         public void TestDictionaryInt()
         {
@@ -923,6 +1020,9 @@ Tax: {{ 7 | match_tax }}";
 
                 var result = Template.Parse("{{a.property_a").Render(context);
                 Assert.AreEqual("A", result);
+
+                result = Template.Parse("{{null_ref?.property_a").Render(context);
+                Assert.AreEqual(string.Empty, result);
 
                 Assert.Catch<ScriptRuntimeException>(() =>
                    Template.Parse("{{a.property_a.null_ref}}").Render(context));