+++ src/Scriban/Runtime/DynamicCustomFunction.cs
@@ -75,12 +75,11 @@ namespace Scriban.Runtime
                 var lastParam = Parameters[Parameters.Length - 1];
                 if (lastParam.ParameterType.IsArray)
                 {
-                    foreach (var param in lastParam.GetCustomAttributes(typeof(ParamArrayAttribute), false))
+                    if (lastParam.ParameterType == typeof(object[]) || lastParam.ParameterType == typeof(ScriptExpression[]) || lastParam.GetCustomAttributes(typeof(ParamArrayAttribute), false).Length > 0)
                     {
                         _varParamKind = ScriptVarParamKind.LastParameter;
                         _paramsElementType = lastParam.ParameterType.GetElementType();
                         _paramsIndex = Parameters.Length - 1;
-                        break;
                     }
                 }
             }