+++ changelog.md
@@ -1,5 +1,25 @@
 # Changelog
 
+> This changelog is no longer used for newer version. Please visits https://github.com/scriban/scriban/releases
+
+## 5.4.0 (6 Feb 2022)
+- Fix async when accessing variables (#382)
+- Fix doc for object.kind (#394)
+- Fix comparison between ulong and int (#395)
+- Add a notice for html.strip that it can fail/timeout in case of malformed HTML (#406)
+- Log an error instead of an exception in an invalid syntax with for/if/else (#393)
+
+## 5.3.0 (6 Feb 2022)
+- Add support for auto-indent (#398)
+
+## 5.2.0 (6 Feb 2022)
+- Add support for specifying custom format and culture for date.parse
+
+## 5.1.0 (5 Feb 2022)
+- Fix accessing null?.property when EnableRelaxedMemberAccess == False (#401)
+- Migrate tests to net6.0 (#410)
+- Add support for anykind of expressions in when clauses (#409)
+
 ## 5.0.0 (20 Oct 2021)
 - Fix scope for variable (global vs local) (#385)